+++ services/financial-service/src/main/java/dev/econolyze/application/dto/TransactionDTO.java
@@ -21,6 +21,6 @@ public class TransactionDTO {
     private String description;
     private LocalDate date;
     private String method;
-    private Long financialGoalId;
+    private Integer financialGoalId;
 
 }
